style(katana-db): standardize naming for historical tables (#1773)
Standardize the naming convention of historical-data related tables.

- `*ChangeHistory` 

tables that store data changes based on the block number it occured.

- `*ChangeSet` 

tables that maintain a **set** of block numbers used for indexing the respective `*ChangeHistory` table. The elements in the set corresponds to the _primary key_ of `*ChangeHistory` table where an entry exist with a _subkey_ that is equal to _primary key_ corresponds to the aforementioned **set**.