katana: update default predeployed accounts (#2524)
ref #2498 #2405 

- update the default predeployed account class implementation to OZ `0.17.0` account [preset](https://github.com/OpenZeppelin/cairo-contracts/blob/v0.17.0/packages/presets/src/account.cairo)
- rename class constants

is it required to change the account implementation because the version we're using doesn't support transaction version > 1 as seen below:

https://github.com/OpenZeppelin/cairo-contracts/blob/6ab91b5cf57d6a7d02f2d5c5abfc7b5712bc8e47/src/account/account.cairo#L75C1-L77C14

technically v3 support is later added in OZ `0.8.1`. so going all the way to `0.17.0` may seem exaggerated, but `0.8.1` is ~8 months old now so better to just use the latest compatible version. tbh i dont exactly know how different `0.17.0` is compared to `0.8.1`.