fix(provider): use RW transaction only when on remote fallback (#9)
Prevent deadlocking on trying to create fork StateProvider due to the fact that the TxValidator maintains a long term reference to the underlying state provider. If we use RW transaction for the state provider, this will cause deadlock as we can't create any new state provider because there can only be a single database RW transaction at a time.