feat(explorer): serve explorer from the RPC server (#15)
Currently, `katana` spawns a dedicated server for the explorer app. While this is fine it doesn't work well when it is from a `slot`-hosted instance. Because the `katana` RPC url injected upon startup directly into the html file of the explorer, as seen here: 

https://github.com/dojoengine/katana/blob/904ee2f1b51641cab48e4dc8ffae076b4cdb2a87/crates/explorer/src/lib.rs#L157-L190

The RPC url however, would be a localhost but `slot` instance, the node would be placed behind a reverse proxy. So, we need a way to forward the information about the `slot` url (ie `https://api.cartridge.gg/x/<NAME>/katana`) to the explorer app.

The approach that I've taken is, instead of spawning a dedicated server, the app is served from the same server as the JSON-RPC under the path `/explorer` (eg `localhost:5050/explorer`).

This requires a corresponding change on the explorer code as well, where a new environment variable is defined to specify if the app is running in an 'embedded' mode.

https://github.com/cartridge-gg/explorer/blob/6e49070599603c8a5f3212cb86ca2c709f131fe8/src/constants/rpc.ts#L14-L19

If it is running in 'embedded' mode, then the base url of the app will be implied as the JSON-RPC url to fetch the blockchain data from.

For example, if you run:

```
katana --explorer --http.port 5050
```

the explorer will be served at `http://localhost:5050/explorer` and `http://localhost:5050` is used as the JSON-RPC url (this is done on the explorer app).