fix(db): enum variants ordering (#142)
This PR fixes wrong enum ordering for the v6 transactions. The ordering matters here as it will affect how the variant is encoded when using binary serialization format like `postcard`.

Changing the order will change how postcard de/serialize the enum:

```rust
// V6 before:
pub enum DeclareTx {
    V0(transaction::DeclareTxV0),  // Position 0
    V1(transaction::DeclareTxV1),  // Position 1  
    V2(transaction::DeclareTxV2),  // Position 2
    V3(DeclareTxV3),              // Position 3
}

// V6 after:
pub enum DeclareTx {
    V1(transaction::DeclareTxV1),  // Position 0
    V2(transaction::DeclareTxV2),  // Position 1
    V3(DeclareTxV3),              // Position 2
    V0(transaction::DeclareTxV0),  // Position 3
}
```

Currently, when reading a `DeclareTx` from the database where the actual variant is a `DeclareTx::V3`, because we're deserializing the bytes as `DeclareTx` (**before**), the entry would end up being deserialized as a `DeclareTx::V2`.

In addition to this, to maintain consistency I decided to rearrange the enum variants for current (v7) `DeclareTx` - positioning V0 as the first variant. This does break the database format but it's fine because we're still in pre-release stage of V7 db format. Thus, this change will be part of the changes introduces by the new db format.